Move Input and OutputData classes to use_case package
