+++ src/main/java/entity/NoteFactoryInterface.java
@@ -1,0 +1,4 @@
+package entity;
+
+public class NoteFactoryInterface {
+}
