+++ src/main/java/entity/localEvent/LocalEventFactory.java
@@ -1,0 +1,4 @@
+package entity.localEvent;
+
+public class LocalEventFactory {
+}
