+++ src/main/java/entity/note/NoteFactory.java
@@ -1,0 +1,4 @@
+package entity.note;
+
+public class NoteFactory {
+}
