+++ src/main/java/use_case/localEvent/LocalEventOutputData.java
@@ -1,4 +1,6 @@
-package OutputData;
+package use_case.localEvent;
+
+import use_case.note.NoteOutputData;
 
 import java.time.LocalDate;
